Minor typo

+++ src/main/scala/org/broadinstitute/dsde/workbench/leonardo/dao/google/HttpGoogleDataprocDAO.scala
@@ -91,7 +91,7 @@ class HttpGoogleDataprocDAO(appName: String,
       ()
     } {
       case e: HttpResponseException if e.getStatusCode == StatusCodes.NotFound.intValue => ()
-      case e: GoogleJsonResponseException if e.getStatusCode == StatusCodes.BadRequest &&
+      case e: GoogleJsonResponseException if e.getStatusCode == StatusCodes.BadRequest.intValue &&
         e.getDetails.getMessage.contains("it has other pending delete operations against it") => ()
     }
   }
